+++ src/apps/stable/routes/asyncRoutes/user.ts
@@ -1,4 +1,4 @@
-import { AsyncRoute } from '../../AsyncRoute';
+import { AsyncRoute } from '../../../../components/router/AsyncRoute';
 
 export const ASYNC_USER_ROUTES: AsyncRoute[] = [
     { path: 'search.html', page: 'search' }
