--- src/routes/index.tsx
@@ -1,2 +1,0 @@
-export * from './appRoutes';
-export * from './experimentalAppRoutes';
